Bugfix for original applications that did not have database

+++ flet_mvc/model.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import sessionmaker
 
 
 class FletMVCModel:
-    def __init__(self, record_type):  # TODO: add hinting
+    def __init__(self, record_type=None):  # TODO: add hinting
         """
         Constructor to set the parent app which will be filled out when the MVC module is added to the app.
         It also sets the 2 other components in the MVC pattern.  That will be filled out when the 3 are combined
@@ -21,24 +21,39 @@ class FletMVCModel:
         self.__connection_string = connection_string
 
     def all(self):
+        if self.__record_type is None:
+            return None
+
         with self.__get_session() as session:
             return session.query(self.__record_type).all()
 
     def create(self, record):
+        if self.__record_type is None:
+            return
+
         with self.__get_session() as session:
             session.add(record)
             session.commit()
 
     def read(self, item_id: int):
+        if self.__record_type is None:
+            return None
+
         with self.__get_session() as session:
             return session.get(self.__record_type, item_id)
 
     def update(self, record):
+        if self.__record_type is None:
+            return
+
         with self.__get_session() as session:
             session.merge(record)
             session.commit()
 
     def delete(self, record):
+        if self.__record_type is None:
+            return
+
         with self.__get_session() as session:
             session.delete(record)
             session.commit()
